Commenting for Cool Cat 1st Draft
Finished commenting my code.
In future, I will also track better when I did in the editor itself and inlcude it in code commenting in a section.

Also, at the bottom, I included code that did not work with my explinations of why I tried it and how I understand it. I think it is also worth looking at.